+++ src/Components/Editor/Context.tsx
@@ -15,7 +15,8 @@ export const EditorContext = React.createContext<{
     models: EditorModel[];
     setModels: React.Dispatch<React.SetStateAction<EditorModel[]>>;
     selectedModel: EditorModel | null;
-    setSelectedModel: React.Dispatch<React.SetStateAction<EditorModel | null>>;
+    selectModel: (model: EditorModel | null) => void;
+    //setSelectedModel: React.Dispatch<React.SetStateAction<EditorModel | null>>;
     camTargetZ: number;
     setCamTargetZ: React.Dispatch<React.SetStateAction<number>>;
     minShade: number;
@@ -37,7 +38,6 @@ export function ContextComponent(props: { children: React.ReactNode }) {
     const [selection, setSelection] = React.useState(new GL.SelectionManager());
     const [models, setModels] = React.useState<EditorModel[]>([]);
     const [selectedModel, setSelectedModel] = React.useState<EditorModel | null>(null);
-    console.log('rendering context');
 
     const [camTargetZ, setCamTargetZ] = React.useState<number>(0);
     const [minShade, setMinShade] = React.useState<number>(0);
@@ -45,6 +45,39 @@ export function ContextComponent(props: { children: React.ReactNode }) {
     const [gridVisible, setGridVisible] = React.useState<boolean>(true);
     const [globalShift, setGlobalShift] = React.useState<vec3 | null>(null);
     const [loadingStatus, setLoadingStatus] = React.useState<string>('');
+
+    const selectModel = (model: EditorModel | null) => {
+        setSelectedModel((prev) => {
+            if (prev !== null && prev !== model) prev.selected = false;
+            if (model !== null && prev !== model) model.selected = true;
+            if (model === null || prev !== model) selection.clearSelection();
+            return model;
+        });
+    };
+
+    React.useEffect(() => {
+        const onChange = () => {
+            const copy = scene.objects.filter((obj) => obj instanceof EditorModel) as EditorModel[];
+            setModels(copy);
+            if (selectedModel !== null && !copy.includes(selectedModel)) setSelectedModel(null);
+        };
+
+        scene.onChange = onChange;
+
+        return () => {
+            scene.removeChange = onChange;
+        };
+    }, [scene, selectedModel]);
+
+    React.useEffect(() => {
+        renderer.onInit = () => {
+            const controls = renderer.window.controls;
+            controls.onPick = (m: GL.Pickable) => {
+                const model = m as EditorModel;
+                selectModel(model);
+            };
+        };
+    }, [renderer]);
 
     return (
         <EditorContext.Provider
@@ -58,7 +91,7 @@ export function ContextComponent(props: { children: React.ReactNode }) {
                 models,
                 setModels,
                 selectedModel,
-                setSelectedModel,
+                selectModel,
                 camTargetZ,
                 setCamTargetZ,
                 minShade,
