+++ src/Components/Editor/SidePanel/Actions/Vitals.tsx
