+++ src/Components/Editor/SidePanel/ViewSettings/Widgets/Shading.tsx
@@ -3,7 +3,7 @@ import { MdHeight } from 'react-icons/md';
 
 import { EditorModel } from '@utils/models/models/EditorModel';
 
-import { EditorContext } from '@components/Editor/Utils/Context';
+import { EditorContext } from '@components/Editor/Context';
 
 import { Range } from '@elements/Range';
 import { DetailWidget, WidgetDescription, WidgetLine, WidgetTitle } from '@elements/Widgets';
