+++ src/Components/Editor/ViewControls.tsx
@@ -6,12 +6,10 @@ import { ShaderControls } from './ViewControls/Shader';
 
 export function ViewControls() {
     return (
-        <>
-            <div className="absolute m-4 space-x-2 left-0 top-0 z-40 flex flex-row ">
-                <ProjectionControls />
-                <DirectionControls />
-                <ShaderControls />
-            </div>
-        </>
+        <div className="absolute m-4 space-x-2 left-0 top-0 z-40 flex flex-row ">
+            <ProjectionControls />
+            <DirectionControls />
+            <ShaderControls />
+        </div>
     );
 }
